third_party/WebKit/LayoutTests/http/tests/htmlimports/import-dynamic-block-async.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
var t = async_test('Dynamically added import blocks following import.');

function followerReady() {
  t.step(function() {
      assert_true(null != dynamicLink.import, 'Dynamic import'); 
      assert_true(null != followingLink.import, 'Following import'); 
      t.done();
  });
}

(function() {
    var link = document.createElement('link');
    link.id = 'dynamicLink'
    link.rel = 'import';
    link.href = 'resources/import-slow-child.cgi';
    document.head.appendChild(link);
})();
</script>
<link async id="followingLink" rel="import" href="resources/hello.html" onload="followerReady()">
</body>
</html>
